refactored and reorganized the code a little bit attempted to make camera follow player, failed miserably

+++ SpaceCommandes/src/com/rauma/lille/stages/ControllerStage.java
@@ -17,7 +17,7 @@ public class ControllerStage extends AbstractStage {
 	private Touchpad touchpadLeft;
 	private Touchpad touchpadRight;
 
-	public ControllerStage(int width, int height, boolean keepAspectRatio) {
+	public ControllerStage(float width, float height, boolean keepAspectRatio) {
 		super(width, height, keepAspectRatio);
 		init();
 	}
